+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - improved polytrope IC
 - optimized refresh when render parameters change
 - Raymarcher: fixed shading artifacts when shadows are enabled
+- optional frame interpolation in animation node
 
 ## 2021-07-18
 - optimized search radius in AsymmetricSolver